lnd+funding: move Manager to funding, change references
